lab 3 program(book details)
Create a class Book which contains four members: name, author, price,
num_pages. Include a constructor to set the values for the members. Include
methods to set and get the details of the objects. Include a toString( ) method
that could display the complete details of the book. Develop a Java program to
create n book objects.